clear component ProductPage + AllSalePage craft 2 homemade hooks with filter logic
